refactor(frontend): decompose StockDetailPage.tsx into StockDetailPage/ directory
Split 694-LOC monolith into co-located panel components under 300 LOC each:
ChartPanel (chart + controls + market profile + extended hours),
MetadataPanel (news feed + trader plan checklist),
ScannerHistoryPanel (event history + force scan dialog).
Shell index.tsx retains all queries, mutations, and state.
